+++ composeApp/src/commonMain/kotlin/com/gs/wialonlocal/features/monitoring/data/entity/Lmsg.kt
@@ -4,12 +4,12 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Lmsg(
-    val f: Int,
+    val f: Long,
     val lc: Long,
     val p: P? = null,
     val pos: Pos? = null,
     val rt: Int,
-    val t: Int,
+    val t: Long,
     val tp: String
 )
